[alarmecoder] Move device state init to handler superclass
